axxx apv : model & business TODOs are complete enough added spring validation added project, improved whole hibernate mapping & model for this switched to session in view (also detransacted services & switched init)

+++ easysoa-samples/easysoa-samples-axxx/axxx-dps-apv/axxx-dps-apv-core/src/main/java/com/axxx/dps/apv/persistence/GenericEntityDao.java
@@ -30,9 +30,9 @@ public interface GenericEntityDao<T extends GenericEntity<T>> {
     public abstract List<T> list(Class<? extends T> objectClass, Criterion filter, Order order, Integer limit,
             Integer offset);
 
-    public abstract Integer count();
+    public abstract Long count();
 
-    public abstract Integer count(Class<? extends T> objectClass, Criterion filter, Order order, Integer limit,
+    public abstract Long count(Class<? extends T> objectClass, Criterion filter, Order order, Integer limit,
             Integer offset);
 
     public abstract void evict(T object);
